chore: update gitignore
We now ignore the contents of data_raw and all contents of data intermediate EXCEPT the downsampled version of the aggregated data and its prepared version with calculated fields for the analysis. The reason for this is so people can download the repo and run the most important scripts without having to manually download the data from the Divvy database.